Added some more classes to check. LOOK at some commentaries. LOOK at Includes
