+++ RefactoredClasses/RefactoredClasses/RefactoredClasses.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="Obstacles.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="PlayerCar.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Traffic.hpp">
@@ -33,6 +36,9 @@
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="Obstacles.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="PlayerCar.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
